hotfix/credential_PhD_removed_on_occupancy_updates
Any updates to `occupancy` were removing the credential "PhD" from the occupancy.  This is because the uppercase version of the credential, i.e. `PHD`(see line 206) was not matching the actual abbreviation stored in the database (PhD).  SQL updated to be case insensitive. This was only affecting the credential PhD, as it is the only credential with a mixed-case abbreviation.